* int => uint16_t for planenum in face_t

+++ src/tools/ufo2map/map.h
@@ -51,7 +51,7 @@ typedef struct face_s {
 
 	struct portal_s	*portal;
 	int				texinfo;
-	int				planenum;
+	uint16_t		planenum;
 	uint32_t		contentFlags;	/**< faces in different contents can't merge */
 	winding_t		*w;
 	int				numpoints;
